move arena to sgx directory
