+++ enclave/core/sgx/arena.h
