+++ enclave/core/sgx/switchlesscalls.c
@@ -10,6 +10,7 @@
 #include <openenclave/internal/utils.h>
 #include "handle_ecall.h"
 #include "platform_t.h"
+#include "arena.h"
 
 // The number of host thread workers. Initialized by host through ECALL
 static size_t _host_worker_count = 0;
@@ -277,4 +278,24 @@ oe_sgx_sleep_switchless_worker_ocall(oe_enclave_worker_context_t* context)
 {
     OE_UNUSED(context);
     return OE_UNSUPPORTED;
+}
+
+// Function used by oeedger8r for allocating switchless ocall buffers.
+// Preallocate a pool of shared memory per thread for switchless ocalls
+// and then allocate memory from that pool. Since OE does not support
+// reentrant ecalls in the same thread, there can at most be one ecall
+// and one ocall active in a thread. Although an enclave function can
+// make multiple OCALLs, the OCALLs are serialized. So the allocation
+// for one OCALL doesn't interfere with the allocation for the next OCALL.
+// A stack-based allocation scheme is the most efficient in this case.
+void* oe_allocate_switchless_ocall_buffer(size_t size)
+{
+    return oe_arena_malloc(size);
+}
+
+// Function used by oeedger8r for freeing ocall buffers.
+void oe_free_switchless_ocall_buffer(void* buffer)
+{
+    OE_UNUSED(buffer);
+    oe_arena_free_all();
 }
